Simplify sofarpc instrumentation: merge modules, add REST/Triple server spans, rpc.method tag
