+++ dd-java-agent/instrumentation/sofarpc/sofarpc-5.0/build.gradle
@@ -11,12 +11,25 @@ apply from: "$rootDir/gradle/java.gradle"
 
 addTestSuiteForDir('latestDepTest', 'test')
 
+configurations.testRuntimeClasspath {
+  resolutionStrategy.force "com.google.guava:guava:32.1.3-jre"
+}
+
 dependencies {
   compileOnly group: "com.alipay.sofa", name: "sofa-rpc-all", version: "5.6.0"
 
-  testImplementation group: "com.alipay.sofa", name: "sofa-rpc-all", version: "5.6.0"
+  testImplementation group: "com.alipay.sofa", name: "sofa-rpc-all", version: "5.14.2"
   // JAX-RS annotations required for the REST protocol test interface (@Path, @GET, etc.)
   testImplementation group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.1.1"
+  // Required so that GrpcServerModule / GrpcClientModule are discovered via ServiceLoader
+  // in SofaRpcTripleWithGrpcForkedTest.
+  testImplementation project(':dd-java-agent:instrumentation:grpc-1.5')
+  testImplementation group: "io.grpc", name: "grpc-netty", version: "1.53.0"
+  testImplementation group: "io.grpc", name: "grpc-core", version: "1.53.0"
+  testImplementation group: "io.grpc", name: "grpc-stub", version: "1.53.0"
+  testImplementation group: "com.google.protobuf", name: "protobuf-java", version: "3.25.3"
+  testImplementation group: "com.alibaba", name: "fastjson", version: "1.2.83"
+  testImplementation group: "com.google.guava", name: "guava", version: "32.1.3-jre"
 
   latestDepTestImplementation group: "com.alipay.sofa", name: "sofa-rpc-all", version: "+"
 }
