+++ dd-java-agent/instrumentation/sofarpc/sofarpc-5.0/src/main/java/datadog/trace/instrumentation/sofarpc/ProviderProxyInvokerInstrumentation.java
@@ -46,17 +46,11 @@ public class ProviderProxyInvokerInstrumentation
       if (protocol == null) {
         return null;
       }
-      // Prefer the parent context stored by transport-specific instrumentation (e.g.
-      // TripleServerInstrumentation reads it from gRPC Metadata). For Bolt and H2C the
-      // transport instrumentations only set the protocol, so parentContext is null here and
-      // we fall back to extracting from SofaRequest.requestProps as before.
-      AgentSpanContext parentContext = SofaRpcProtocolContext.getParentContext();
-      if (parentContext == null) {
-        parentContext = extractContextAndGetSpanContext(request, GETTER);
-      }
-      // parentContext may be null for Triple+gRPC-enabled: TripleServerInstrumentation skips
-      // Metadata extraction when a grpc.server span is already active. In that case
-      // startSpan() without an explicit parent naturally attaches to the active grpc.server span.
+      // For Bolt and H2C the client injects trace context into SofaRequest.requestProps;
+      // extract it here. For Triple, parentContext will be null and startSpan() without an
+      // explicit parent naturally attaches to the active grpc.server span. For REST,
+      // parentContext will also be null and the active netty.request span becomes the parent.
+      AgentSpanContext parentContext = extractContextAndGetSpanContext(request, GETTER);
       AgentSpan span =
           parentContext != null
               ? startSpan(SOFA_RPC_SERVER, parentContext)
