+++ dd-java-agent/instrumentation/sofarpc/sofarpc-5.0/src/main/java/datadog/trace/instrumentation/sofarpc/SofaRpcModule.java
@@ -31,6 +31,8 @@ public class SofaRpcModule extends InstrumenterModule.Tracing {
         new AbstractClusterInstrumentation(),
         new BoltServerProcessorInstrumentation(),
         new H2cServerTaskInstrumentation(),
+        new RestServerHandlerInstrumentation(),
+        new TripleServerInstrumentation(),
         new ProviderProxyInvokerInstrumentation());
   }
 }
