+++ dd-java-agent/instrumentation/sofarpc/sofarpc-5.0/src/main/java/datadog/trace/instrumentation/sofarpc/TripleServerInstrumentation.java
@@ -2,29 +2,34 @@ package datadog.trace.instrumentation.sofarpc;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.takesNoArguments;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import datadog.trace.agent.tooling.Instrumenter;
 import net.bytebuddy.asm.Advice;
 
-public class H2cServerTaskInstrumentation
+public class TripleServerInstrumentation
     implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
 
   @Override
   public String instrumentedType() {
-    return "com.alipay.sofa.rpc.server.http.AbstractHttpServerTask";
+    return "com.alipay.sofa.rpc.server.triple.UniqueIdInvoker";
   }
 
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
-        isMethod().and(named("run")).and(takesNoArguments()), getClass().getName() + "$RunAdvice");
+        isMethod()
+            .and(named("invoke"))
+            .and(takesArguments(1))
+            .and(takesArgument(0, named("com.alipay.sofa.rpc.core.request.SofaRequest"))),
+        getClass().getName() + "$InvokeAdvice");
   }
 
-  public static class RunAdvice {
+  public static class InvokeAdvice {
     @Advice.OnMethodEnter(suppress = Throwable.class)
     public static void enter() {
-      SofaRpcProtocolContext.set("h2c");
+      SofaRpcProtocolContext.set("tri");
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
