+++ dd-java-agent/instrumentation/sofarpc/sofarpc-5.0/src/test/groovy/datadog/trace/instrumentation/sofarpc/SofaRpcRestTest.groovy
@@ -21,11 +21,11 @@ import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
 /**
  * Tests SOFA RPC REST protocol instrumentation.
  *
- * REST works out-of-the-box via dd-trace-java's Apache HttpClient and HTTP server
- * instrumentation.  Our SOFA RPC instrumentation contributes the sofarpc.request[client]
- * span (from AbstractClusterInstrumentation).  There is no sofarpc.request[server] span
- * because no transport-level instrumentation sets SofaRpcProtocolContext for REST — the
- * server side is covered by JAX-RS / HTTP server instrumentation instead.
+ * Our instrumentation contributes sofarpc.request[client] (AbstractClusterInstrumentation)
+ * and sofarpc.request[server] (RestServerHandlerInstrumentation + ProviderProxyInvokerInstrumentation).
+ * Distributed trace propagation is delegated to dd-trace-java's HTTP instrumentation
+ * (Apache HttpClient on the client side, Netty on the server side), which is not active in
+ * this unit test — so the server span appears as a separate trace root here.
  */
 class SofaRpcRestTest extends InstrumentationSpecification {
 
@@ -41,36 +41,36 @@ class SofaRpcRestTest extends InstrumentationSpecification {
   def setupSpec() {
     ServerConfig serverConfig =
       new ServerConfig()
-        .setProtocol("rest")
-        .setHost("127.0.0.1")
-        .setPort(port)
+      .setProtocol("rest")
+      .setHost("127.0.0.1")
+      .setPort(port)
 
     ProviderConfig<GreeterService> providerConfig =
       new ProviderConfig<GreeterService>()
-        .setApplication(new ApplicationConfig().setAppName("test-server"))
-        .setInterfaceId(GreeterService.name)
-        .setRef(new GreeterServiceImpl())
-        .setServer(serverConfig)
-        .setRegister(false)
+      .setApplication(new ApplicationConfig().setAppName("test-server"))
+      .setInterfaceId(GreeterService.name)
+      .setRef(new GreeterServiceImpl())
+      .setServer(serverConfig)
+      .setRegister(false)
 
     restProviderBootstrap = providerConfig.export()
 
     greeterService =
       new ConsumerConfig<GreeterService>()
-        .setApplication(new ApplicationConfig().setAppName("test-client"))
-        .setInterfaceId(GreeterService.name)
-        .setDirectUrl("rest://127.0.0.1:${port}")
-        .setProtocol("rest")
-        .setRegister(false)
-        .setSubscribe(false)
-        .refer()
+      .setApplication(new ApplicationConfig().setAppName("test-client"))
+      .setInterfaceId(GreeterService.name)
+      .setDirectUrl("rest://127.0.0.1:${port}")
+      .setProtocol("rest")
+      .setRegister(false)
+      .setSubscribe(false)
+      .refer()
   }
 
   def cleanupSpec() {
     restProviderBootstrap?.unExport()
   }
 
-  def "client span created for REST call"() {
+  def "client and server spans created for REST call"() {
     setup:
     String serviceUniqueName = GreeterService.name + ":1.0"
 
@@ -81,11 +81,8 @@ class SofaRpcRestTest extends InstrumentationSpecification {
     reply == "Hello, World"
 
     and:
-    // REST is instrumented by Apache HttpClient (client) + HTTP server (server) integrations.
-    // Our SOFA RPC instrumentation adds sofarpc.request[client] on top.  No server-side
-    // SofaRpcProtocolContext is set for REST, so ProviderProxyInvoker produces no sofarpc span.
-    // Only the client-side trace is collected here.
-    assertTraces(1) {
+    assertTraces(2) {
+      // trace(0): client side — caller + sofarpc.request[client]
       trace(2) {
         basicSpan(it, "caller")
         span {
@@ -96,12 +93,34 @@ class SofaRpcRestTest extends InstrumentationSpecification {
           childOf span(0)
           tags {
             "$Tags.RPC_SERVICE" serviceUniqueName
+            "rpc.method" "sayHello"
             "rpc.system" "sofarpc"
             "sofarpc.protocol" "rest"
             "component" "sofarpc-client"
             "span.kind" "client"
             peerServiceFrom(Tags.RPC_SERVICE)
             defaultTags()
+          }
+        }
+      }
+      // trace(1): server side — sofarpc.request[server].
+      // SofaRequest.getTargetServiceUniqueName() is null on the server side for REST
+      // (not propagated through the JAX-RS layer), so resourceName is the method name only
+      // and rpc.service tag is absent. Parent link to the client trace is provided by
+      // HTTP instrumentation (not active in this test), so this span is a trace root here.
+      trace(1) {
+        span {
+          operationName "sofarpc.request"
+          resourceName "sayHello"
+          spanType "rpc"
+          errored false
+          tags {
+            "rpc.method" "sayHello"
+            "rpc.system" "sofarpc"
+            "sofarpc.protocol" "rest"
+            "component" "sofarpc-server"
+            "span.kind" "server"
+            defaultTags(true)
           }
         }
       }
