--- dd-java-agent/instrumentation/sofarpc/sofarpc-5.14/build.gradle
@@ -1,47 +1,0 @@
-muzzle {
-  pass {
-    group = "com.alipay.sofa"
-    module = "sofa-rpc-all"
-    versions = "[5.14.2,)"
-    assertInverse = true
-  }
-}
-
-apply from: "$rootDir/gradle/java.gradle"
-
-configurations.testRuntimeClasspath {
-  // Force the JRE variant of Guava so that java.time-based APIs (e.g.
-  // CacheBuilder.expireAfterWrite(Duration)) are available at test runtime.
-  // Without this, dependency resolution picks the android variant which lacks
-  // those overloads, causing NoSuchMethodError during SOFA RPC initialisation.
-  resolutionStrategy.force "com.google.guava:guava:32.1.3-jre"
-}
-
-addTestSuiteForDir('latestDepTest', 'test')
-
-dependencies {
-  compileOnly project(':dd-java-agent:instrumentation:sofarpc:sofarpc-5.0')
-  compileOnly group: "com.alipay.sofa", name: "sofa-rpc-all", version: "5.14.2"
-  // grpc-api provides io.grpc.Metadata used by TripleGrpcMetadataExtractAdapter.
-  compileOnly group: "io.grpc", name: "grpc-api", version: "1.53.0"
-
-  testImplementation project(':dd-java-agent:instrumentation:sofarpc:sofarpc-5.0')
-  // Required so that GrpcServerModule / GrpcClientModule are discovered via ServiceLoader
-  // in SofaRpcTripleWithGrpcForkedTest (the test framework loads InstrumenterModules from
-  // the test classpath, not from a shadow jar).
-  testImplementation project(':dd-java-agent:instrumentation:grpc-1.5')
-  testImplementation group: "com.alipay.sofa", name: "sofa-rpc-all", version: "5.14.2"
-  // sofa-rpc-all:5.14.2 was compiled against gRPC 1.53.0 and does not bundle its
-  // transitive dependencies — they must be declared explicitly at the matching version.
-  testImplementation group: "io.grpc", name: "grpc-netty", version: "1.53.0"
-  testImplementation group: "io.grpc", name: "grpc-core", version: "1.53.0"
-  testImplementation group: "io.grpc", name: "grpc-stub", version: "1.53.0"
-  testImplementation group: "com.google.protobuf", name: "protobuf-java", version: "3.25.3"
-  testImplementation group: "com.alibaba", name: "fastjson", version: "1.2.83"
-  // sofa-common-tools (transitive via sofa-rpc-all:5.14.2) uses Guava 28+ API
-  // (CacheBuilder.expireAfterWrite(Duration)); force the JRE variant — the android
-  // variant omits java.time-based overloads that are required at runtime.
-  testImplementation group: "com.google.guava", name: "guava", version: "32.1.3-jre"
-
-  latestDepTestImplementation group: "com.alipay.sofa", name: "sofa-rpc-all", version: "+"
-}
