--- dd-java-agent/instrumentation/sofarpc/sofarpc-5.14/src/main/java/datadog/trace/instrumentation/sofarpc/SofaRpcTripleModule.java
@@ -1,27 +1,0 @@
-package datadog.trace.instrumentation.sofarpc;
-
-import com.google.auto.service.AutoService;
-import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.agent.tooling.InstrumenterModule;
-import java.util.Collections;
-import java.util.List;
-
-@AutoService(InstrumenterModule.class)
-public class SofaRpcTripleModule extends InstrumenterModule.Tracing {
-
-  public SofaRpcTripleModule() {
-    super("sofarpc");
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      packageName + ".TripleGrpcMetadataExtractAdapter",
-    };
-  }
-
-  @Override
-  public List<Instrumenter> typeInstrumentations() {
-    return Collections.singletonList(new TripleServerInstrumentation());
-  }
-}
